requirements.txt
fastapi
uvicorn
pillow
pytesseract
pandas
unstructured
python-multipart
